requirements.txt
beautifulsoup4==4.8.2
jupyterlab==2.1.0
kaggle==1.5.6
notebook==6.0.3
pandas==1.0.3
pip==20.0.2
pyprojroot==0.2.0
pyyaml==5.3.1
soupsieve==2.0
